requirement.txt
streamlit
tensorflow
scikit-learn
joblib
pillow
numpy
